requirements.txt
certifi==2018.8.13
chardet==3.0.4
Django==2.1
djangorestframework==3.8.2
enum34==1.1.6
future==0.16.0
idna==2.7
num2words==0.5.7
numpy==1.14.0
plac==0.9.6
python-crfsuite==0.9.6
python-decouple==3.1
pytz==2018.5
requests==2.19.1
scikit-learn==0.19.2
scipy==1.1.0
semantic-version==2.6.0
six==1.11.0
sklearn-crfsuite==0.3.6
snips-nlu==0.16.2
snips-nlu-es==0.2.0
snips-nlu-ontology==0.57.3
snips-nlu-utils==0.6.1
tabulate==0.8.2
tqdm==4.25.0
urllib3==1.23
